requirements.txt
jupyter
nb_black
numpy
pandas
scikit-learn
matplotlib
seaborn
altair
uszipcode
tqdm
lxml
fake-useragent
beautifulsoup4
selenium
tables
bokeh
dask[distributed]
dask[dataframe]
dask[delayed]
dask-ml[complete]
fsspec>=0.3.3
wordninja
pyldavis
spacy
nltk
corextopic
gensim
networkx
requests
goose3
newspaper3k
fastapi
uvicorn
python-multipart
azure-storage-blob
